Add new lobpcg integration test and move Marchenko-Pastur into it
